requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-Bootstrap==3.3.0.1
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
ORM==1.1.0
SQLAlchemy==0.9.8
WTForms==2.0.1
Werkzeug==0.9.6
backports.pbkdf2==0.1
inflection==0.2.1
itsdangerous==0.24
webassets==0.10.1
wtf==0.1
